requirements.txt
asgiref==3.8.1
click==8.2.1
colorama==0.4.6
Django==5.2.1
django-crispy-forms==2.4
django-rq==3.0.1
et_xmlfile==2.0.0
openpyxl==3.1.5
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
redis==6.2.0
rq==2.3.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
